backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
azure-common==1.1.28
azure-communication-email==1.0.0
azure-core==1.39.0
azure-mgmt-core==1.6.0
azure-storage-blob==12.28.0
azure-storage-file-datalake==12.17.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.135.3
fastexcel==0.12.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.4
Mako==1.3.10
MarkupSafe==3.0.3
msrest==0.7.1
oauthlib==3.3.1
openpyxl==3.1.5
packaging==26.0
pluggy==1.6.0
polars==1.38.1
polars-runtime-32==1.38.1
psycopg2-binary==2.9.11
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pytest==8.3.3
python-dotenv==1.2.2
requests==2.33.1
requests-oauthlib==2.0.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
